Refactor and support multi-query in map.html
This refactors the javascript in map.html a bit (for better or for worse
I do not know), but it should allow the browser to release some objects.

It also modifies the link update routine to use the multi-query mode in the
backend so the browser only makes 1 GET request instead of 34 per second.